feat: Canonical Type + Function Mapping (Phase 2)
- 方言内置 type_to_canonical + canonical_to_type 映射表
- 方言内置 function_to_canonical + canonical_to_function 映射表
- 删除 type_mapping.py (30条规则 → 映射引擎)
- Pipeline 新增 Canonical 类型映射和函数映射阶段
- Canonical 映射在所有 Rule 阶段之后运行，避免误匹配
- 结构变换函数保留为 Rule (constraint 阶段)
- 规则数: 53 → 25

+++ tests/actual_sniff_oracle.sql
@@ -11,7 +11,7 @@ DROP TABLE IF EXISTS "test_oracle" CASCADE;
 -- 步骤2: 表结构
 -- --------------------------------------------
 CREATE TABLE "test_oracle" (
-  "id" INTEGER NOT NULL,
+  "id" NUMERIC NOT NULL,
   "desc" TEXT
 );
